requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor-5==0.2.10
django-cleanup==8.0.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-recaptcha==3.0.0
django-taggit==5.0.1
Pillow==10.1.0
psycopg2==2.9.9
pytils==0.4.1
sqlparse==0.4.4
tzdata==2023.3
